v1.9.7

+++ vms/platformvm/state/state.go
@@ -139,8 +139,8 @@ type State interface {
 	// Commit changes to the base database.
 	Commit() error
 
-	// Returns a batch of unwritten changes that, when written, will commit all
-	// pending changes to the base database.
+	// Returns a batch of unwritten changes that, when written, will be commit
+	// all pending changes to the base database.
 	CommitBatch() (database.Batch, error)
 
 	Close() error
@@ -228,10 +228,8 @@ type state struct {
 	currentHeight uint64
 
 	addedBlocks map[ids.ID]stateBlk // map of blockID -> Block
-	// cache of blockID -> Block
-	// If the block isn't known, nil is cached.
-	blockCache cache.Cacher[ids.ID, *stateBlk]
-	blockDB    database.Database
+	blockCache  cache.Cacher        // cache of blockID -> Block, if the entry is nil, it is not in the database
+	blockDB     database.Database
 
 	validatorsDB                 database.Database
 	currentValidatorsDB          database.Database
@@ -253,18 +251,18 @@ type state struct {
 	pendingSubnetDelegatorBaseDB database.Database
 	pendingSubnetDelegatorList   linkeddb.LinkedDB
 
-	validatorWeightDiffsCache cache.Cacher[string, map[ids.NodeID]*ValidatorWeightDiff] // cache of heightWithSubnet -> map[ids.NodeID]*ValidatorWeightDiff
+	validatorWeightDiffsCache cache.Cacher // cache of heightWithSubnet -> map[ids.NodeID]*ValidatorWeightDiff
 	validatorWeightDiffsDB    database.Database
 
-	validatorPublicKeyDiffsCache cache.Cacher[uint64, map[ids.NodeID]*bls.PublicKey] // cache of height -> map[ids.NodeID]*bls.PublicKey
+	validatorPublicKeyDiffsCache cache.Cacher // cache of height -> map[ids.NodeID]*bls.PublicKey
 	validatorPublicKeyDiffsDB    database.Database
 
-	addedTxs map[ids.ID]*txAndStatus            // map of txID -> {*txs.Tx, Status}
-	txCache  cache.Cacher[ids.ID, *txAndStatus] // txID -> {*txs.Tx, Status}. If the entry is nil, it isn't in the database
+	addedTxs map[ids.ID]*txAndStatus // map of txID -> {*txs.Tx, Status}
+	txCache  cache.Cacher            // cache of txID -> {*txs.Tx, Status} if the entry is nil, it is not in the database
 	txDB     database.Database
 
-	addedRewardUTXOs map[ids.ID][]*avax.UTXO            // map of txID -> []*UTXO
-	rewardUTXOsCache cache.Cacher[ids.ID, []*avax.UTXO] // txID -> []*UTXO
+	addedRewardUTXOs map[ids.ID][]*avax.UTXO // map of txID -> []*UTXO
+	rewardUTXOsCache cache.Cacher            // cache of txID -> []*UTXO
 	rewardUTXODB     database.Database
 
 	modifiedUTXOs map[ids.ID]*avax.UTXO // map of modified UTXOID -> *UTXO if the UTXO is nil, it has been removed
@@ -276,17 +274,17 @@ type state struct {
 	subnetBaseDB  database.Database
 	subnetDB      linkeddb.LinkedDB
 
-	transformedSubnets     map[ids.ID]*txs.Tx            // map of subnetID -> transformSubnetTx
-	transformedSubnetCache cache.Cacher[ids.ID, *txs.Tx] // cache of subnetID -> transformSubnetTx if the entry is nil, it is not in the database
+	transformedSubnets     map[ids.ID]*txs.Tx // map of subnetID -> transformSubnetTx
+	transformedSubnetCache cache.Cacher       // cache of subnetID -> transformSubnetTx if the entry is nil, it is not in the database
 	transformedSubnetDB    database.Database
 
-	modifiedSupplies map[ids.ID]uint64             // map of subnetID -> current supply
-	supplyCache      cache.Cacher[ids.ID, *uint64] // cache of subnetID -> current supply if the entry is nil, it is not in the database
+	modifiedSupplies map[ids.ID]uint64 // map of subnetID -> current supply
+	supplyCache      cache.Cacher      // cache of subnetID -> current supply if the entry is nil, it is not in the database
 	supplyDB         database.Database
 
-	addedChains  map[ids.ID][]*txs.Tx                    // maps subnetID -> the newly added chains to the subnet
-	chainCache   cache.Cacher[ids.ID, []*txs.Tx]         // cache of subnetID -> the chains after all local modifications []*txs.Tx
-	chainDBCache cache.Cacher[ids.ID, linkeddb.LinkedDB] // cache of subnetID -> linkedDB
+	addedChains  map[ids.ID][]*txs.Tx // maps subnetID -> the newly added chains to the subnet
+	chainCache   cache.Cacher         // cache of subnetID -> the chains after all local modifications []*txs.Tx
+	chainDBCache cache.Cacher         // cache of subnetID -> linkedDB
 	chainDB      database.Database
 
 	// The persisted fields represent the current database value
@@ -372,10 +370,10 @@ func new(
 	metricsReg prometheus.Registerer,
 	rewards reward.Calculator,
 ) (*state, error) {
-	blockCache, err := metercacher.New[ids.ID, *stateBlk](
+	blockCache, err := metercacher.New(
 		"block_cache",
 		metricsReg,
-		&cache.LRU[ids.ID, *stateBlk]{Size: blockCacheSize},
+		&cache.LRU{Size: blockCacheSize},
 	)
 	if err != nil {
 		return nil, err
@@ -398,39 +396,39 @@ func new(
 	pendingSubnetDelegatorBaseDB := prefixdb.New(subnetDelegatorPrefix, pendingValidatorsDB)
 
 	validatorWeightDiffsDB := prefixdb.New(validatorWeightDiffsPrefix, validatorsDB)
-	validatorWeightDiffsCache, err := metercacher.New[string, map[ids.NodeID]*ValidatorWeightDiff](
+	validatorWeightDiffsCache, err := metercacher.New(
 		"validator_weight_diffs_cache",
 		metricsReg,
-		&cache.LRU[string, map[ids.NodeID]*ValidatorWeightDiff]{Size: validatorDiffsCacheSize},
+		&cache.LRU{Size: validatorDiffsCacheSize},
 	)
 	if err != nil {
 		return nil, err
 	}
 
 	validatorPublicKeyDiffsDB := prefixdb.New(validatorPublicKeyDiffsPrefix, validatorsDB)
-	validatorPublicKeyDiffsCache, err := metercacher.New[uint64, map[ids.NodeID]*bls.PublicKey](
+	validatorPublicKeyDiffsCache, err := metercacher.New(
 		"validator_pub_key_diffs_cache",
 		metricsReg,
-		&cache.LRU[uint64, map[ids.NodeID]*bls.PublicKey]{Size: validatorDiffsCacheSize},
+		&cache.LRU{Size: validatorDiffsCacheSize},
 	)
 	if err != nil {
 		return nil, err
 	}
 
-	txCache, err := metercacher.New[ids.ID, *txAndStatus](
+	txCache, err := metercacher.New(
 		"tx_cache",
 		metricsReg,
-		&cache.LRU[ids.ID, *txAndStatus]{Size: txCacheSize},
+		&cache.LRU{Size: txCacheSize},
 	)
 	if err != nil {
 		return nil, err
 	}
 
 	rewardUTXODB := prefixdb.New(rewardUTXOsPrefix, baseDB)
-	rewardUTXOsCache, err := metercacher.New[ids.ID, []*avax.UTXO](
+	rewardUTXOsCache, err := metercacher.New(
 		"reward_utxos_cache",
 		metricsReg,
-		&cache.LRU[ids.ID, []*avax.UTXO]{Size: rewardUTXOsCacheSize},
+		&cache.LRU{Size: rewardUTXOsCacheSize},
 	)
 	if err != nil {
 		return nil, err
@@ -444,37 +442,37 @@ func new(
 
 	subnetBaseDB := prefixdb.New(subnetPrefix, baseDB)
 
-	transformedSubnetCache, err := metercacher.New[ids.ID, *txs.Tx](
+	transformedSubnetCache, err := metercacher.New(
 		"transformed_subnet_cache",
 		metricsReg,
-		&cache.LRU[ids.ID, *txs.Tx]{Size: chainCacheSize},
+		&cache.LRU{Size: chainCacheSize},
 	)
 	if err != nil {
 		return nil, err
 	}
 
-	supplyCache, err := metercacher.New[ids.ID, *uint64](
+	supplyCache, err := metercacher.New(
 		"supply_cache",
 		metricsReg,
-		&cache.LRU[ids.ID, *uint64]{Size: chainCacheSize},
+		&cache.LRU{Size: chainCacheSize},
 	)
 	if err != nil {
 		return nil, err
 	}
 
-	chainCache, err := metercacher.New[ids.ID, []*txs.Tx](
+	chainCache, err := metercacher.New(
 		"chain_cache",
 		metricsReg,
-		&cache.LRU[ids.ID, []*txs.Tx]{Size: chainCacheSize},
+		&cache.LRU{Size: chainCacheSize},
 	)
 	if err != nil {
 		return nil, err
 	}
 
-	chainDBCache, err := metercacher.New[ids.ID, linkeddb.LinkedDB](
+	chainDBCache, err := metercacher.New(
 		"chain_db_cache",
 		metricsReg,
-		&cache.LRU[ids.ID, linkeddb.LinkedDB]{Size: chainDBCacheSize},
+		&cache.LRU{Size: chainDBCacheSize},
 	)
 	if err != nil {
 		return nil, err
@@ -658,11 +656,11 @@ func (s *state) GetSubnetTransformation(subnetID ids.ID) (*txs.Tx, error) {
 		return tx, nil
 	}
 
-	if tx, cached := s.transformedSubnetCache.Get(subnetID); cached {
-		if tx == nil {
+	if txIntf, cached := s.transformedSubnetCache.Get(subnetID); cached {
+		if txIntf == nil {
 			return nil, database.ErrNotFound
 		}
-		return tx, nil
+		return txIntf.(*txs.Tx), nil
 	}
 
 	transformSubnetTxID, err := database.GetID(s.transformedSubnetDB, subnetID[:])
@@ -688,8 +686,8 @@ func (s *state) AddSubnetTransformation(transformSubnetTxIntf *txs.Tx) {
 }
 
 func (s *state) GetChains(subnetID ids.ID) ([]*txs.Tx, error) {
-	if chains, cached := s.chainCache.Get(subnetID); cached {
-		return chains, nil
+	if chainsIntf, cached := s.chainCache.Get(subnetID); cached {
+		return chainsIntf.([]*txs.Tx), nil
 	}
 	chainDB := s.getChainDB(subnetID)
 	chainDBIt := chainDB.NewIterator()
@@ -720,15 +718,16 @@ func (s *state) AddChain(createChainTxIntf *txs.Tx) {
 	createChainTx := createChainTxIntf.Unsigned.(*txs.CreateChainTx)
 	subnetID := createChainTx.SubnetID
 	s.addedChains[subnetID] = append(s.addedChains[subnetID], createChainTxIntf)
-	if chains, cached := s.chainCache.Get(subnetID); cached {
+	if chainsIntf, cached := s.chainCache.Get(subnetID); cached {
+		chains := chainsIntf.([]*txs.Tx)
 		chains = append(chains, createChainTxIntf)
 		s.chainCache.Put(subnetID, chains)
 	}
 }
 
 func (s *state) getChainDB(subnetID ids.ID) linkeddb.LinkedDB {
-	if chainDB, cached := s.chainDBCache.Get(subnetID); cached {
-		return chainDB
+	if chainDBIntf, cached := s.chainDBCache.Get(subnetID); cached {
+		return chainDBIntf.(linkeddb.LinkedDB)
 	}
 	rawChainDB := prefixdb.New(subnetID[:], s.chainDB)
 	chainDB := linkeddb.NewDefault(rawChainDB)
@@ -740,10 +739,11 @@ func (s *state) GetTx(txID ids.ID) (*txs.Tx, status.Status, error) {
 	if tx, exists := s.addedTxs[txID]; exists {
 		return tx.tx, tx.status, nil
 	}
-	if tx, cached := s.txCache.Get(txID); cached {
-		if tx == nil {
+	if txIntf, cached := s.txCache.Get(txID); cached {
+		if txIntf == nil {
 			return nil, status.Unknown, database.ErrNotFound
 		}
+		tx := txIntf.(*txAndStatus)
 		return tx.tx, tx.status, nil
 	}
 	txBytes, err := s.txDB.Get(txID[:])
@@ -785,7 +785,7 @@ func (s *state) GetRewardUTXOs(txID ids.ID) ([]*avax.UTXO, error) {
 		return utxos, nil
 	}
 	if utxos, exists := s.rewardUTXOsCache.Get(txID); exists {
-		return utxos, nil
+		return utxos.([]*avax.UTXO), nil
 	}
 
 	rawTxDB := prefixdb.New(txID[:], s.rewardUTXODB)
@@ -869,12 +869,12 @@ func (s *state) GetCurrentSupply(subnetID ids.ID) (uint64, error) {
 		return supply, nil
 	}
 
-	cachedSupply, ok := s.supplyCache.Get(subnetID)
+	supplyIntf, ok := s.supplyCache.Get(subnetID)
 	if ok {
-		if cachedSupply == nil {
+		if supplyIntf == nil {
 			return 0, database.ErrNotFound
 		}
-		return *cachedSupply, nil
+		return supplyIntf.(uint64), nil
 	}
 
 	supply, err := database.GetUInt64(s.supplyDB, subnetID[:])
@@ -886,7 +886,7 @@ func (s *state) GetCurrentSupply(subnetID ids.ID) (uint64, error) {
 		return 0, err
 	}
 
-	s.supplyCache.Put(subnetID, &supply)
+	s.supplyCache.Put(subnetID, supply)
 	return supply, nil
 }
 
@@ -929,8 +929,8 @@ func (s *state) GetValidatorWeightDiffs(height uint64, subnetID ids.ID) (map[ids
 	}
 	prefixStr := string(prefixBytes)
 
-	if weightDiffs, ok := s.validatorWeightDiffsCache.Get(prefixStr); ok {
-		return weightDiffs, nil
+	if weightDiffsIntf, ok := s.validatorWeightDiffsCache.Get(prefixStr); ok {
+		return weightDiffsIntf.(map[ids.NodeID]*ValidatorWeightDiff), nil
 	}
 
 	rawDiffDB := prefixdb.New(prefixBytes, s.validatorWeightDiffsDB)
@@ -959,8 +959,8 @@ func (s *state) GetValidatorWeightDiffs(height uint64, subnetID ids.ID) (map[ids
 }
 
 func (s *state) GetValidatorPublicKeyDiffs(height uint64) (map[ids.NodeID]*bls.PublicKey, error) {
-	if publicKeyDiffs, ok := s.validatorPublicKeyDiffsCache.Get(height); ok {
-		return publicKeyDiffs, nil
+	if publicKeyDiffsIntf, ok := s.validatorPublicKeyDiffsCache.Get(height); ok {
+		return publicKeyDiffsIntf.(map[ids.NodeID]*bls.PublicKey), nil
 	}
 
 	heightBytes := database.PackUInt64(height)
@@ -1228,7 +1228,7 @@ func (s *state) loadCurrentValidators() error {
 
 			validator := s.currentStakers.getOrCreateValidator(staker.SubnetID, staker.NodeID)
 			if validator.delegators == nil {
-				validator.delegators = btree.NewG(defaultTreeDegree, (*Staker).Less)
+				validator.delegators = btree.New(defaultTreeDegree)
 			}
 			validator.delegators.ReplaceOrInsert(staker)
 
@@ -1314,7 +1314,7 @@ func (s *state) loadPendingValidators() error {
 
 			validator := s.pendingStakers.getOrCreateValidator(staker.SubnetID, staker.NodeID)
 			if validator.delegators == nil {
-				validator.delegators = btree.NewG(defaultTreeDegree, (*Staker).Less)
+				validator.delegators = btree.New(defaultTreeDegree)
 			}
 			validator.delegators.ReplaceOrInsert(staker)
 
@@ -1513,7 +1513,7 @@ func (s *state) writeBlocks() error {
 		}
 
 		delete(s.addedBlocks, blkID)
-		s.blockCache.Put(blkID, &stBlk)
+		s.blockCache.Put(blkID, stateBlk)
 		if err := s.blockDB.Put(blkID[:], blockBytes); err != nil {
 			return fmt.Errorf("failed to write block %s: %w", blkID, err)
 		}
@@ -1522,13 +1522,15 @@ func (s *state) writeBlocks() error {
 }
 
 func (s *state) GetStatelessBlock(blockID ids.ID) (blocks.Block, choices.Status, error) {
-	if blk, ok := s.addedBlocks[blockID]; ok {
+	if blk, exists := s.addedBlocks[blockID]; exists {
 		return blk.Blk, blk.Status, nil
 	}
-	if blkState, ok := s.blockCache.Get(blockID); ok {
-		if blkState == nil {
-			return nil, choices.Processing, database.ErrNotFound
+	if blkIntf, cached := s.blockCache.Get(blockID); cached {
+		if blkIntf == nil {
+			return nil, choices.Processing, database.ErrNotFound // status does not matter here
 		}
+
+		blkState := blkIntf.(stateBlk)
 		return blkState.Blk, blkState.Status, nil
 	}
 
@@ -1551,7 +1553,7 @@ func (s *state) GetStatelessBlock(blockID ids.ID) (blocks.Block, choices.Status,
 		return nil, choices.Processing, err
 	}
 
-	s.blockCache.Put(blockID, &blkState)
+	s.blockCache.Put(blockID, blkState)
 	return blkState.Blk, blkState.Status, nil
 }
 
@@ -1590,55 +1592,58 @@ func (s *state) writeCurrentStakers(updateValidators bool, height uint64) error 
 			// Copy [nodeID] so it doesn't get overwritten next iteration.
 			nodeID := nodeID
 
-			weightDiff := &ValidatorWeightDiff{
-				Decrease: validatorDiff.validatorDeleted,
-			}
-			switch {
-			case validatorDiff.validatorAdded:
+			var (
+				weightDiff     = &ValidatorWeightDiff{}
+				isNewValidator bool
+			)
+			if validatorDiff.validatorModified {
+				// This validator is being added or removed.
 				staker := validatorDiff.validator
+
+				weightDiff.Decrease = validatorDiff.validatorDeleted
 				weightDiff.Amount = staker.Weight
 
-				// The validator is being added.
-				vdr := &uptimeAndReward{
-					txID:        staker.TxID,
-					lastUpdated: staker.StartTime,
+				if validatorDiff.validatorDeleted {
+					// Invariant: Only the Primary Network contains non-nil
+					//            public keys.
+					if staker.PublicKey != nil {
+						// Record the public key of the validator being removed.
+						pkDiffs[nodeID] = staker.PublicKey
 
-					UpDuration:      0,
-					LastUpdated:     uint64(staker.StartTime.Unix()),
-					PotentialReward: staker.PotentialReward,
-				}
-
-				vdrBytes, err := blocks.GenesisCodec.Marshal(blocks.Version, vdr)
-				if err != nil {
-					return fmt.Errorf("failed to serialize current validator: %w", err)
-				}
-
-				if err = validatorDB.Put(staker.TxID[:], vdrBytes); err != nil {
-					return fmt.Errorf("failed to write current validator to list: %w", err)
-				}
-
-				s.validatorUptimes.LoadUptime(nodeID, subnetID, vdr)
-			case validatorDiff.validatorDeleted:
-				staker := validatorDiff.validator
-				weightDiff.Amount = staker.Weight
-
-				// Invariant: Only the Primary Network contains non-nil
-				//            public keys.
-				if staker.PublicKey != nil {
-					// Record the public key of the validator being removed.
-					pkDiffs[nodeID] = staker.PublicKey
-
-					pkBytes := bls.PublicKeyToBytes(staker.PublicKey)
-					if err := pkDiffDB.Put(nodeID[:], pkBytes); err != nil {
-						return err
+						pkBytes := bls.PublicKeyToBytes(staker.PublicKey)
+						if err := pkDiffDB.Put(nodeID[:], pkBytes); err != nil {
+							return err
+						}
 					}
-				}
 
-				if err := validatorDB.Delete(staker.TxID[:]); err != nil {
-					return fmt.Errorf("failed to delete current staker: %w", err)
-				}
+					if err := validatorDB.Delete(staker.TxID[:]); err != nil {
+						return fmt.Errorf("failed to delete current staker: %w", err)
+					}
 
-				s.validatorUptimes.DeleteUptime(nodeID, subnetID)
+					s.validatorUptimes.DeleteUptime(nodeID, subnetID)
+				} else {
+					// The validator is being added.
+					vdr := &uptimeAndReward{
+						txID:        staker.TxID,
+						lastUpdated: staker.StartTime,
+
+						UpDuration:      0,
+						LastUpdated:     uint64(staker.StartTime.Unix()),
+						PotentialReward: staker.PotentialReward,
+					}
+
+					vdrBytes, err := blocks.GenesisCodec.Marshal(blocks.Version, vdr)
+					if err != nil {
+						return fmt.Errorf("failed to serialize current validator: %w", err)
+					}
+
+					if err = validatorDB.Put(staker.TxID[:], vdrBytes); err != nil {
+						return fmt.Errorf("failed to write current validator to list: %w", err)
+					}
+
+					s.validatorUptimes.LoadUptime(nodeID, subnetID, vdr)
+					isNewValidator = true
+				}
 			}
 
 			err := writeCurrentDelegatorDiff(
@@ -1678,7 +1683,7 @@ func (s *state) writeCurrentStakers(updateValidators bool, height uint64) error 
 			if weightDiff.Decrease {
 				err = validators.RemoveWeight(s.cfg.Validators, subnetID, nodeID, weightDiff.Amount)
 			} else {
-				if validatorDiff.validatorAdded {
+				if isNewValidator {
 					staker := validatorDiff.validator
 					err = validators.Add(
 						s.cfg.Validators,
@@ -1776,16 +1781,17 @@ func writePendingDiff(
 	pendingDelegatorList linkeddb.LinkedDB,
 	validatorDiff *diffValidator,
 ) error {
-	if validatorDiff.validatorAdded {
-		err := pendingValidatorList.Put(validatorDiff.validator.TxID[:], nil)
-		if err != nil {
-			return fmt.Errorf("failed to add pending validator: %w", err)
+	if validatorDiff.validatorModified {
+		staker := validatorDiff.validator
+
+		var err error
+		if validatorDiff.validatorDeleted {
+			err = pendingValidatorList.Delete(staker.TxID[:])
+		} else {
+			err = pendingValidatorList.Put(staker.TxID[:], nil)
 		}
-	}
-	if validatorDiff.validatorDeleted {
-		err := pendingValidatorList.Delete(validatorDiff.validator.TxID[:])
 		if err != nil {
-			return fmt.Errorf("failed to delete pending validator: %w", err)
+			return fmt.Errorf("failed to update pending validator: %w", err)
 		}
 	}
 
@@ -1897,9 +1903,8 @@ func (s *state) writeTransformedSubnets() error {
 
 func (s *state) writeSubnetSupplies() error {
 	for subnetID, supply := range s.modifiedSupplies {
-		supply := supply
 		delete(s.modifiedSupplies, subnetID)
-		s.supplyCache.Put(subnetID, &supply)
+		s.supplyCache.Put(subnetID, supply)
 		if err := database.PutUInt64(s.supplyDB, subnetID[:], supply); err != nil {
 			return fmt.Errorf("failed to write subnet supply: %w", err)
 		}
